Revert "restructuring (#89)"
This reverts commit 877dd81acb9dca3b5b492b81025180f7977bcb22.
