+++ src/students/example/ml/model.py
@@ -1,0 +1,11 @@
+def add_one(number):
+    return number + 1 + 3
+
+
+def main():
+    result = add_one(1)
+    print(result)
+
+
+if __name__ == '__main__':
+    main()
